Python/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
et_xmlfile==2.0.0
idna==3.11
json5==0.12.1
keyboard==0.13.5
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
psycopg2==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-snap7==2.0.2
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Unipath==1.1
urllib3==2.5.0
